Adds timeout fetch control to lib
